+++ testapps/GH944/src/main/java/test/JiayouOrderTestEntity.java
@@ -1,0 +1,7 @@
+package test;
+
+import act.util.SimpleBean;
+
+public class JiayouOrderTestEntity extends BaseModel implements SimpleBean {
+    public String name;
+}